pyproject.toml
[project]
name = "openeo_mountains_snow"
version = "0.1.0"
description = "Snow processing in the mountains using openEO"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openeo",
    "hydra-core"
]

[project.optional-dependencies]
dev = [
    "pytest"
]

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
pythonpath = [
  "src"
]